# contracts/openapi/core-api.v1.yaml
openapi: 3.1.0
info:
  title: Urban Crisis Management — Core API
  version: 1.0.0
  description: |
    REST contract of the core service (modular monolith). Conventions (detailed design §6.1):
    - Bearer token (OIDC; local dev tokens only with APP_ENV=local). Authorization = permission + organisational scope,
      resolved server-side; deny-by-default. Hidden resources answer 404.
    - `Idempotency-Key` is REQUIRED on retryable creates (reports, incidents, alerts, dispatch). Same key + same body
      replays the stored response (`Idempotent-Replayed: true`); same key + different body → 422 IDEMPOTENCY_KEY_REUSED.
    - Every response carries `X-Correlation-ID`. Errors use the fixed envelope `ErrorEnvelope`.
    - Mutations of versioned aggregates require `version` (optimistic locking → 409 VERSION_CONFLICT).
    - Consequential decisions (reject, transition, cancel, allocation, approval) require a `reason` and are audited.
    - Times are RFC 3339 UTC. Coordinates are WGS84 (EPSG:4326) decimal degrees.
servers:
  - url: /
...
security:
  - bearer: []
paths:
...
  /api/v1/media/blob:
    get:
      tags: [media]
      security: []
      summary: Signed download (filesystem store only)
      parameters:
        - {name: k, in: query, required: true, schema: {type: string}}
        - {name: exp, in: query, required: true, schema: {type: integer}}
        - {name: sig, in: query, required: true, schema: {type: string}}
      responses: {"200": {description: file}, "403": {$ref: "#/components/responses/Error"}}
...
components:
  securitySchemes:
    bearer: {type: http, scheme: bearer, bearerFormat: JWT}
...
  responses:
    Error:
      description: Error in the fixed envelope
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, correlation_id]
          properties:
            code: {type: string, examples: [VALIDATION_ERROR, FORBIDDEN, NOT_FOUND, VERSION_CONFLICT, IDEMPOTENCY_KEY_REUSED, NOT_APPROVED, FOUR_EYES_REQUIRED, RESOURCE_UNAVAILABLE, RATE_LIMITED]}
            message: {type: string}
            correlation_id: {type: string}
            details: {type: [array, "null"], items: {type: object, properties: {field: {type: string}, reason: {type: string}}}}
...